Add toggle between dark and light color tmux/vim
Helps on toggling from a dark theme to a light theme in tmux and vim.

The theme being used (light or dark) is kept in the tmux theme
environment variable.

From the terminal switching between light or dark is done with `ol` or
`od`. In vim the `leader o` helps setting the color vim theme used in
tmux.